% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesmiss.R
\name{bayesmiss}
\alias{bayesmiss}
\title{Bayesian regression with missing covariates}
\usage{
bayesmiss(originaldata, smtype, smformula, method, order)
}
\arguments{
\item{originaldata}{the data frame upon which the analysis is to be performed.}

\item{smtype}{a string specifying the substantive model type}

\item{smformula}{an expression for the substantive model type}

\item{method}{a vector of strings specifying the imputation method for each of
the columns in \code{originaldata}. Currently the possible values are "norm"
(normal linear model), "logit" (logistic regression), "mlogit" (multinomial
logistic regression, and "ologit" (ordinal logistic regression". The elements
corresponding to fully observed variables should be the empty string "".}

\item{order}{a vector specifying the order in which the joint model for missing
covariates should be constructed. e.g. c(0,0,1,2) specifies that the first two
variables are fully observed and that the third and fourth are partially observed,
with the covariate model constructed by modelling the marginal distribution of
the third variable and the conditional distribution of the fourth given the third.}
}
\value{
\code{bayesmiss} generates two files in the current working directory:
\code{bayesmissmod.bug} is a JAGS model file for the constructed model, and \code{bayesmissRscript.r}
is an R file containing R code for generating the required JAGS parameters and data
objects, and a call to the \code{jags} function of the \code{R2jags} package for
fitting the model.
}
\description{
\code{bayesmiss} generates JAGS model code and an R script to perform
Bayesian regression, allowing for missingness in covariates.
}
\details{
\code{bayesmiss} faciliates running Bayesian regression models in which there
are missing values in some of the covariates. The function generates two files

The \code{order} argument is used to specify the order in which the joint
distribution of the partially observed covariates is to be modelled. Models
for the partially observed covariates automatically condition on substantive model
covariates which are fully observed.

The R script file generated contains commands to generate the required JAGS data
and parameter objects, which are passed when calling \code{jags}. Note that
\code{bayesmiss} doesn't actually run this code - the user, having ensured that
the \code{R2jags} package is installed, must run this code to fit the model.
Note that the MCMC options in the call to \code{jags} are just suggested default.
It is up to the user to ensure, via the usual diagnostics for MCMC, that a
sufficient number of iterations have been run to ensure convergence of the chains.

If it is desired to add interactions or non-linear covariate effects, the easiest
approach is to first run \code{bayesmiss} omitting these terms, and then modify
the JAGS code file and R code specifying the priors as needed.
}

